[MOD-16145] tests: move shared timeout/shard helpers to common.py
Replace the cross-test-file import (`from test_blocked_client_timeout
import ...`) added in the previous commit with proper shared helpers in
common.py. Importing one test module from another runs its collection-time
code and couples the files; helpers used by more than one test file belong
in common.py.

- Move pid_cmd, non_coord_shard_conns, split_shards_pick_one_paused,
  assert_timeout_warning and ON_TIMEOUT_CONFIG into common.py.
- Promote split_shards_pick_one_paused and non_coord_shard_conns to public
  names (drop the leading underscore): both are called across the
  module boundary, and `from common import *` does not export _-prefixed
  names (common.py defines no __all__).
- Dedupe pid_cmd: it was copy-pasted in test_blocked_client_timeout.py and
  test_query_oom.py; both now use the common.py definition.
- test_hybrid_timeout.py drops the cross-file import; the symbols come from
  its existing `from common import *`.

Pure relocation/rename, no behavior change.
